arch-vega: Pass s_memtime through smem pipe
Previously, the cycle count was written with no latency.
Now, the cycle count is written when the scalar memory pipeline executes s_memtime, because a s_memtime should be treated like a s_load_dwordx2.
Through microbenchmarks, the s_memtime latency has been set to 41 clock cycles.
Memtimes are marked hasNoAddr so that they can pass through to the VIPERSequencer,
if other instructions are added that use hasNoAddr their behavior should be added to hitCallback

Change-Id: I39a8c56990fdc0054be00a1effe16f06a3da4836