mem: WIP change does not compile
Directory and TCC changes to correctly implement WB L2 caches.
This change tries to fix the issue with WB L2 caches in FS mode simulation.
There is currently an error with the wt_sendResponseWBAckTrig action.
The action copies the w_sendResponseWBAck action and peeks into the
Trigger queue instead of the responseNetwork queue.
This has issues because we directly cannot use the same fields
that response message type provides as they are not all available
in the trigger message type. This change does not allow you to compile.

Change-Id: I1e46549399c23bbcbd3f3ea1a398902aa68ee9ca